jdbcTemplate

+++ src/main/java/io/pivotal/pal/tracker/TimeEntry.java
@@ -5,13 +5,13 @@ import java.util.Objects;
 
 public class TimeEntry {
     private long id;
-    private int projectId;
-    private int userId;
+    private long projectId;
+    private long userId;
     private LocalDate date;
     private int hours;
 
 
-    public TimeEntry(long id, int projectId, int userId, LocalDate date, int hours) {
+    public TimeEntry(long id, long projectId, long userId, LocalDate date, int hours) {
         this.id = id;
         this.projectId = projectId;
         this.userId = userId;
@@ -20,7 +20,7 @@ public class TimeEntry {
 
     }
 
-    public TimeEntry(int projectId, int userId, LocalDate date, int hours) {
+    public TimeEntry(long projectId, long userId, LocalDate date, int hours) {
 
         this.projectId = projectId;
         this.userId = userId;
@@ -43,19 +43,19 @@ public class TimeEntry {
         this.id = id;
     }
 
-    public int getProjectId() {
+    public long getProjectId() {
         return projectId;
     }
 
-    public void setProjectId(int projectId) {
+    public void setProjectId(long projectId) {
         this.projectId = projectId;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
